% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smle_resid.R
\name{smle_resid}
\alias{smle_resid}
\title{Extract residuals of the full likelihood proportional hazards model}
\usage{
smle_resid(y, d, x, fit, type = c("score", "deviance"))
}
\arguments{
\item{y}{survival time (> 0).}

\item{d}{right-censoring indicator, \code{1}: observed; \code{0}: right-censored.}

\item{x}{p-dimensional covariates matrix.}

\item{fit}{an object comes from the function \code{smle_ph}.}

\item{type}{type of residual, either \code{deviance} or \code{score}.}
}
\value{
\code{smle_resid} returns a numeric vector (if \code{type = "deviance"}) or a matrix (if \code{type = "score"}) of residuals extracted from the \code{object}.
}
\description{
This function extracts residuals of the full likelihood proportional hazards model estimated by the sieve estimation. Deviance-type and score-type residuals are available.
}
\details{
see Halabi et al., (2024+) for detailed method explanation.
}
\examples{
library(smlePH)
# The 'fit' comes from an example description of smle_ph()
smle_resid(y = y, d = d, x = x, fit = fit, type = "deviance")
smle_resid(y = y, d = d, x = x, fit = fit, type = "score")
}
\references{
Halabi et al., (2024+) Sieve maximum full likelihood estimation for the proportional hazards model
}
