% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDensRatioGOF.R
\name{testDensRatioGOF}
\alias{testDensRatioGOF}
\title{Goodness-of-Fit Test of the Validity of a Univariate or Multivariate Mark Density Ratio Model}
\usage{
testDensRatioGOF(mark, tx, theta = NULL, lambda = NULL, B = 1000)
}
\arguments{
\item{mark}{either a numeric vector specifying a univariate continuous mark or a matrix specifying a multivariate continuous mark, 
completely observed in all cases (i.e., failures). No missing mark values are permitted.}

\item{tx}{a numeric vector indicating the treatment group (1 if treatment, 0 if placebo) in all cases. No missing values are permitted.}

\item{theta}{a numeric vector of the coefficients \eqn{\phi} in the weight function \eqn{g(v, \phi)} in the density ratio model}

\item{lambda}{the Lagrange multiplier in the profile score functions for \eqn{\phi} (that arises by profiling out the nuisance parameter)}

\item{B}{the number of bootstrap iterations (1000 by default)}
}
\description{
\code{testDensRatioGoF} implements the complete-case goodness-of-fit test of Qin and Zhang (1997) for the validity of the mark density ratio model in 
Juraska and Gilbert (2013). \code{testDensRatioGoF} can handle univariate and multivariate marks, but the mark data must be fully observed in all failures.
}
\examples{
n <- 500
tx <- rep(0:1, each=n/2)
tm <- c(rexp(n/2, 0.1), rexp(n/2, 0.1 * exp(-0.4)))
cens <- runif(n, 0, 15)
eventTime <- pmin(tm, cens, 3)
eventInd <- as.numeric(tm <= pmin(cens, 3))
mark1 <- ifelse(eventInd==1, c(rbeta(n/2, 2, 5), rbeta(n/2, 2, 2)), NA)
mark2 <- ifelse(eventInd==1, c(rbeta(n/2, 1, 3), rbeta(n/2, 5, 1)), NA)

mark1 <- mark1[eventInd==1]
mark2 <- mark2[eventInd==1]
tx <- tx[eventInd==1]

# perform the test with a univariate mark
testDensRatioGOF(mark1, tx, B=20)

# perform the test with a bivariate mark
testDensRatioGOF(cbind(mark1, mark2), tx, B=20)

}
\references{
Juraska, M. and Gilbert, P. B. (2013), Mark-specific hazard ratio model with multivariate continuous marks: an application to vaccine efficacy. \emph{Biometrics} 69(2):328-337.

Qin, J., & Zhang, B. (1997). A goodness-of-fit test for logistic regression models based on case-control data. \emph{Biometrika}, 84(3), 609-618.
}
